Works without error handling

+++ controllers/blogs.js
@@ -9,12 +9,8 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const body = request.body
+  const blog = new Blog(request.body)
 
-  // If title and url are both missing a bad request error is sent
-  if(!body.title && !body.url){
-    return response.status(400).json({ error: 'url and title are missing' })
-  }
   if (!request.token) {
     return response.status(401).json({ error: 'token missing' })
   }
@@ -24,17 +20,21 @@ blogsRouter.post('/', async (request, response) => {
     // before jwt.verify
     return response.status(401).json({ error: 'invalid token' })
   }
+  if(!blog.title || !blog.url){
+    return response.status(400).json({ error: 'url and title are required' })
+  }
   const user = await User.findById(decodedToken.id)
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user._id
-  })
+  if(!blog.likes) {
+    blog.likes = 0
+  }
+
+  blog.user = user
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
@@ -43,17 +43,34 @@ blogsRouter.delete('/:id', async (request, response) => {
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const deletedBlog = await Blog.findOneAndRemove({ _id: request.params.id, user: decodedToken.id })
-  if(deletedBlog){
-    response.status(204).end()
-  }else {
-    response.status(400).json({ error: 'no such blog entry for this user' })
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'only the creator can delete blogs' })
   }
-})
+  await Blog.findOneAndRemove({ _id: request.params.id, user: decodedToken.id })
+  user.blogs = user.blogs.filter(b => b.id.toString() !== request.params.id.toString())
+  await user.save()
+  response.status(204).end()
 
-blogsRouter.patch('/:id', async (request, response) => {
+})
+blogsRouter.put('/:id', async (request, response) => {
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id,request.body,{ new: true })
+  if(!updatedBlog){
+    return response.status(404).end()
+  }
   response.status(200).json(updatedBlog)
+})
+blogsRouter.post('/:id/comments', async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  if(!blog){
+    return response.status(404).end()
+  }
+  const { comment } = request.body
+  blog.comments = [...blog.comments,{ text: comment }]
+  const savedBlog = await blog.save()
+  response.status(201).send(savedBlog)
 })
 
 module.exports = blogsRouter
